network-rpc: Refactor the rpc api && client to a new crate
