+++ block-relayer/Cargo.toml
@@ -27,7 +27,8 @@ starcoin-block-relayer-api = { path = "./api"}
 starcoin-canonical-serialization = { package="starcoin-canonical-serialization", path = "../commons/scs"}
 starcoin-types = {path = "../types", package = "starcoin-types" }
 starcoin-metrics = {path = "../commons/metrics"}
-network-rpc = { path = "../network-rpc", package = "starcoin-network-rpc" }
+starcoin-network-rpc-api = { path = "../network-rpc/api" }
+
 [dev-dependencies]
 tokio = { version = "0.2", features = ["full"] }
 actix-rt = "1.1"
