+++ block-relayer/src/block_relayer.rs
@@ -4,9 +4,9 @@ use bus::Bus;
 use bus::{Broadcast, BusActor, Subscription};
 use crypto::HashValue;
 use logger::prelude::*;
-use network_rpc::{gen_client::NetworkRpcClient, GetTxns};
 use starcoin_block_relayer_api::{NetCmpctBlockMessage, PeerCmpctBlockEvent};
 use starcoin_network::network::NetworkAsyncService;
+use starcoin_network_rpc_api::{gen_client::NetworkRpcClient, GetTxns};
 use starcoin_sync::helper::get_txns;
 use starcoin_sync_api::PeerNewBlock;
 use starcoin_txpool_api::TxPoolSyncService;
