+++ network-rpc/Cargo.toml
@@ -40,6 +40,8 @@ genesis = { path = "../core/genesis", package = "starcoin-genesis" }
 txpool-api = {path = "../txpool/api", package = "starcoin-txpool-api"}
 state_api = { path="../state/api", package = "starcoin-state-api" }
 state-service = { path = "../state/service", package = "starcoin-state-service"}
+starcoin-network-rpc-api = { path= "api" }
+
 [dev-dependencies]
 network = {path = "../network", package = "starcoin-network" }
 consensus = { path = "../consensus", package = "starcoin-consensus" }
