+++ network-rpc/api/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "starcoin-network-rpc-api"
+version = "0.3.1"
+authors = ["Starcoin Core Dev <dev@starcoin.org>"]
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+starcoin-types = { path = "../../types" }
+starcoin-txpool = { path = "../../txpool"}
+network-api = { path="../../network/api" }
+network-rpc-core = { path = "../core"}
+starcoin-accumulator = { path = "../../core/accumulator" }
+starcoin-crypto = { path = "../../commons/crypto" }
+starcoin-state-api = { path="../../state/api" }
+starcoin-state-tree = { path = "../../state/state-tree" }
+network-rpc-derive = { path = "../derive" }
+scs = { path = "../../commons/scs", package = "starcoin-canonical-serialization" }
+starcoin-logger = {path = "../../commons/logger"}
+
+serde = { version = "1.0", features = ["derive"] }
+actix = "0.10.0-alpha.3"
+anyhow = "1.0"
+futures = "0.3"
