+++ network-rpc/core/src/delegates.rs
@@ -1,3 +1,6 @@
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
 use anyhow::Result;
 use futures::prelude::future::BoxFuture;
 use futures::Future;
