+++ network-rpc/core/src/lib.rs
@@ -1,2 +1,5 @@
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
 pub mod delegates;
 pub mod server;
