+++ network-rpc/derive/src/lib.rs
@@ -1,3 +1,6 @@
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
 mod helper;
 mod rpc_trait;
 mod to_client;
