+++ sync/src/state_sync/mod.rs
@@ -18,9 +18,9 @@ use futures::executor::block_on;
 use logger::prelude::*;
 use network::NetworkAsyncService;
 use network_api::NetworkService;
-use network_rpc::gen_client::NetworkRpcClient;
 use starcoin_accumulator::node::{AccumulatorStoreType, ACCUMULATOR_PLACEHOLDER_HASH};
 use starcoin_accumulator::AccumulatorNode;
+use starcoin_network_rpc_api::gen_client::NetworkRpcClient;
 use starcoin_state_tree::StateNode;
 use starcoin_storage::Store;
 use std::collections::{HashMap, VecDeque};
