+++ sync/tests/full_sync_test.rs
@@ -84,7 +84,7 @@ fn test_network_actor_rpc() {
         let state_service_1 =
             ChainStateActor::launch(bus_1.clone(), storage_1.clone(), None).unwrap();
 
-        network_rpc::start_network_rpc_server(
+        starcoin_network_rpc::start_network_rpc_server(
             rx_1,
             first_chain.clone(),
             storage_1.clone(),
@@ -189,7 +189,7 @@ fn test_network_actor_rpc() {
         // network rpc server
         let state_service_2 =
             ChainStateActor::launch(bus_2.clone(), storage_2.clone(), None).unwrap();
-        network_rpc::start_network_rpc_server(
+        starcoin_network_rpc::start_network_rpc_server(
             rx_2,
             second_chain.clone(),
             storage_2.clone(),
