+++ sync/tests/txn_sync_test.rs
@@ -82,7 +82,7 @@ fn test_txn_sync_actor() {
         let state_service_1 =
             ChainStateActor::launch(bus_1.clone(), storage_1.clone(), None).unwrap();
         // network rpc server
-        network_rpc::start_network_rpc_server(
+        starcoin_network_rpc::start_network_rpc_server(
             rpc_rx,
             first_chain.clone(),
             storage_1.clone(),
@@ -168,7 +168,7 @@ fn test_txn_sync_actor() {
         let state_service_2 =
             ChainStateActor::launch(bus_2.clone(), storage_2.clone(), None).unwrap();
         // network rpc server
-        network_rpc::start_network_rpc_server(
+        starcoin_network_rpc::start_network_rpc_server(
             rpc_rx_2,
             second_chain.clone(),
             storage_2.clone(),
